Move `TriplanarTexturesNode` to `triplanarTextures()` Fn
